validateChannelUpdateSignatures in utils
