+++ modules/protocol/src/testing/sync.spec.ts
@@ -14,6 +14,7 @@ import {
   MemoryMessagingService,
   getTestLoggers,
   createTestChannelUpdate,
+  validateChannelUpdateSignatures,
 } from "@connext/vector-utils";
 import {
   UpdateType,
@@ -32,7 +33,6 @@ import Sinon from "sinon";
 import { VectorChainReader } from "@connext/vector-contracts";
 
 // Import as full module for easy sinon function mocking
-import * as vectorUtils from "../utils";
 import * as vectorValidation from "../validate";
 import { inbound, outbound } from "../sync";
 
@@ -549,7 +549,7 @@ describe("outbound", () => {
     validateAndApplyInboundStub = Sinon.stub(vectorValidation, "validateAndApplyInboundUpdate");
 
     // Stub out all signature validation
-    validateUpdateSignatureStub = Sinon.stub(vectorUtils, "validateChannelUpdateSignatures").resolves(
+    validateUpdateSignatureStub = Sinon.createStubInstance(validateChannelUpdateSignatures).resolves(
       Result.ok(undefined),
     );
   });
